+++ sembast/lib/src/api/records_ref.dart
@@ -1,5 +1,4 @@
 import 'package:sembast/src/api/protected/type.dart';
-import 'package:sembast/src/api/record_ref.dart';
 import 'package:sembast/src/api/store_ref.dart';
 
 ///
@@ -11,9 +10,6 @@ abstract class RecordsRef<K extends Key?, V extends Value?> {
 
   /// Record key, null for new record.
   List<K> get keys;
-
-  /// Record ref at a given index.
-  RecordRef<K, V> operator [](int index);
 
   /// Cast if needed.
   RecordsRef<RK, RV> cast<RK extends Key?, RV extends Value?>();